fixed refresh token check to be before every api call
* added refactoring
