+++ lib/api.js
@@ -4,7 +4,7 @@ const open = require('open');
 const puppeteer = require('puppeteer');
 const { cliInputs, getSpinner } = require('./setup');
 const {
-  AUTH: { SCOPES, STATE },
+  AUTH: { SCOPES, STATE, REFRESH_ACCESS_TOKEN_SECONDS },
   INPUT_TYPES,
   MAX_LIMIT_DEFAULT,
   SERVER: {
@@ -27,8 +27,21 @@ const scopes = [
   SCOPES.USERS_TOP_TRACKS,
 ];
 
+let nextTokenRefreshTime;
+
 module.exports = {
   spotifyApi,
+  verifyCredentials: async function () {
+    if (!nextTokenRefreshTime || (nextTokenRefreshTime < new Date())) {
+      nextTokenRefreshTime = new Date();
+      nextTokenRefreshTime.setSeconds(
+        nextTokenRefreshTime.getSeconds() + REFRESH_ACCESS_TOKEN_SECONDS,
+      );
+      const spinner = getSpinner();
+      spinner.info('Generating new access token');
+      await this.checkCredentials();
+    }
+  },
   checkCredentials: async function () {
     if (await spotifyApi.getRefreshToken()) {
       await this.refreshToken();
@@ -137,6 +150,7 @@ module.exports = {
     return extractedTracks;
   },
   extractTrack: async function (trackId) {
+    await this.verifyCredentials();
     return this.parseTrack((await spotifyApi.getTrack(trackId)).body);
   },
   parseTrack: function (track) {
@@ -150,6 +164,7 @@ module.exports = {
     };
   },
   extractPlaylist: async function (playlistId) {
+    await this.verifyCredentials();
     const data = (await spotifyApi.getPlaylist(
       playlistId,
       { limit: MAX_LIMIT_DEFAULT },
@@ -158,6 +173,7 @@ module.exports = {
     if (data.tracks.next) {
       let offset = tracks.length;
       while (tracks.length < data.tracks.total) {
+        await this.verifyCredentials();
         const playlistTracksData = (await spotifyApi
           .getPlaylistTracks(
             playlistId,
@@ -178,6 +194,7 @@ module.exports = {
     };
   },
   extractAlbum: async function (albumId) {
+    await this.verifyCredentials();
     const data = (await spotifyApi.getAlbum(
       albumId,
       { limit: MAX_LIMIT_DEFAULT },
@@ -186,6 +203,7 @@ module.exports = {
     if (data.tracks.next) {
       let offset = tracks.length;
       while (tracks.length < data.tracks.total) {
+        await this.verifyCredentials();
         const albumTracks = (await spotifyApi
           .getAlbumTracks(
             albumId,
@@ -206,6 +224,7 @@ module.exports = {
     };
   },
   extractArtist: async function (artistId) {
+    await this.verifyCredentials();
     const data = (await spotifyApi.getArtist(artistId)).body;
     return {
       id: data.id,
@@ -214,6 +233,7 @@ module.exports = {
     };
   },
   extractArtistAlbums: async function (artistId) {
+    await this.verifyCredentials();
     const artistAlbums = (await spotifyApi.getArtistAlbums(
       artistId,
       { limit: MAX_LIMIT_DEFAULT },
@@ -222,6 +242,7 @@ module.exports = {
     if (artistAlbums.next) {
       let offset = albums.length;
       while (albums.length < artistAlbums.total) {
+        await this.verifyCredentials();
         const additionalArtistAlbums = (await spotifyApi
           .getArtistAlbums(
             artistId,
@@ -235,6 +256,7 @@ module.exports = {
     return albums.filter(album => album.artists[0].id == artistId);
   },
   extractSavedAlbums: async function () {
+    await this.verifyCredentials();
     const savedAlbums = (await spotifyApi.getMySavedAlbums(
       { limit: MAX_LIMIT_DEFAULT },
     )).body;
@@ -242,6 +264,7 @@ module.exports = {
     if (savedAlbums.next) {
       let offset = albums.length;
       while (albums.length < savedAlbums.total) {
+        await this.verifyCredentials();
         const additionalSavedAlbums = (await spotifyApi
           .getMySavedAlbums(
             { limit: MAX_LIMIT_DEFAULT, offset: offset },
@@ -254,6 +277,7 @@ module.exports = {
     return albums.map(album => album.album);
   },
   extractSavedPlaylists: async function () {
+    await this.verifyCredentials();
     const savedPlaylists = (await spotifyApi.getUserPlaylists(
       { limit: MAX_LIMIT_DEFAULT },
     )).body;
@@ -261,6 +285,7 @@ module.exports = {
     if (savedPlaylists.next) {
       let offset = playlists.length;
       while (playlists.length < savedPlaylists.total) {
+        await this.verifyCredentials();
         const additionalSavedPlaylists = (await spotifyApi
           .getUserPlaylists(
             { limit: MAX_LIMIT_DEFAULT, offset: offset },
@@ -274,6 +299,7 @@ module.exports = {
   },
   extractSavedTracks: async function () {
     const spinner = getSpinner();
+    await this.verifyCredentials();
     const savedTracks = (await spotifyApi.getMySavedTracks(
       { limit: MAX_LIMIT_DEFAULT },
     )).body;
@@ -283,6 +309,7 @@ module.exports = {
     if (savedTracks.next) {
       let offset = tracks.length;
       while (tracks.length < savedTracks.total) {
+        await this.verifyCredentials();
         const additionalSavedTracks = (await spotifyApi
           .getMySavedTracks(
             { limit: MAX_LIMIT_DEFAULT, offset: offset },
